+++ CHANGELOG.md
@@ -45,7 +45,9 @@
 
 ### Improvements
 
+- Draw invisible symbols for invisible operators (U+2061, U+2062, U+2063, and U+2064).
 - Update the Lisp, Lua, Scheme, and Tcl syntaxes to update interpreters (thanks to Renfei!).
+- [non-AppStore ver.] Update Sparkle from 2.6.3 to 2.6.4.
 
 
 ### Fixes