+++ CHANGELOG.md
@@ -24,6 +24,7 @@
 - Remove Solarized themes from the bundle.
 - Update all the bundled themes to have a 70% opacity in the current line highlight.
 - Improve the performance of counting values in the editor for the status bar and the document inspector to avoid flicking.
+- Set the editor's kerning to zero when a monospaced font is set via the font panel.
 - Make more table columns sortable.
 - [trivial] Organize the structure of the Edit menu.
 - [trivial] Suppress display of the “Extracting” message on the navigation bar in instantaneous parsing.