+++ CHANGELOG.md
@@ -26,6 +26,7 @@
 - Update all the bundled themes to have a 70% opacity in the current line highlight.
 - Improve the performance of counting values in the editor for the status bar and the document inspector to avoid flicking.
 - Set the editor's kerning to zero when a monospaced font is set via the font panel.
+- Avoid loading errors on themes when missing value exists.
 - Make more table columns sortable.
 - [trivial] Organize the structure of the Edit menu.
 - [trivial] Suppress display of the “Extracting” message on the navigation bar in instantaneous parsing.