+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Change Log
 
-5.0.0 (unreleased)
+5.0.0-beta (661)
 --------------------------
 
 ### New Features